use float instead of int for cpu

+++ src/edu/illinois/cs446/HardwareMonitor.java
@@ -27,7 +27,7 @@ public class HardwareMonitor {
 					lastLine = line;
 				lastLine = lastLine.trim();
 				int start = lastLine.lastIndexOf(" ") + 1;
-				cpuUsage = 100 - Integer.parseInt(lastLine.substring(start));
+				cpuUsage = 100 - (int)Float.parseFloat(lastLine.substring(start));
 				System.out.println(cpuUsage);
 			} catch (IOException e) {
 				e.printStackTrace();
